LayoutTests/transitions/border-radius-transition.html

<!DOCTYPE>

<html lang="en">
<head>
  <style type="text/css" media="screen">
    .box {
      height: 100px;
      width: 100px;
      margin: 10px;
      border: 3px solid blue;
      -webkit-transition-duration: 0.5s;
      -webkit-transition-timing-function: linear;
    }
    
    #box {
      -webkit-transition-property: border-radius;
    }
    
    #box.final {
        border-top-left-radius: 50px;
    }

    #box2 {
      -webkit-transition-property: -webkit-border-radius;
    }
    
    #box2.final {
        -webkit-border-top-left-radius: 50px;
    }
    
  </style>
  <script src="transition-test-helpers.js" type="text/javascript" charset="utf-8"></script>
  <script type="text/javascript" charset="utf-8">

    const expectedValues = [
      // [time, element-id, property, expected-value, tolerance]
      [0.25, 'box', 'border-top-left-radius', 25, 2],
      [0.25, 'box2', '-webkit-border-top-left-radius', 25, 2],
    ];
  
    function setupTest()
    {
      document.getElementById('box').className = 'box final';
      document.getElementById('box2').className = 'box final';
    }
  
    runTransitionTest(expectedValues, setupTest, dontUsePauseAPI); // the pause API does not work with shorthands.
  </script>
</head>
<body>

  <div id="box" class="box"></div>
  <div id="box2" class="box"></div>

  <div id="result">
  </div>

</body>
</html>
